requirements.txt
python-dotenv>=1.0.1
pandas>=2.2.2
requests>=2.32.0
requests-html>=0.10.0
lxml[html_clean]
